requirements.txt
pytest==8.3.1
allure-pytest==2.13.5
allure-python-commons==2.13.5

selenium==4.23.0
webdriver-manager==4.0.2

requests==2.32.3

certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.2

iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
attrs==23.2.0
colorama==0.4.6
typing_extensions==4.9.0
